Some more thinking, experimenting.
The README lays out a lot of sequence diagrams.

It is clear this can be done with a queue, and that a lot of this can be
async, with lots of workers pulling from queues.

Minimal state in the living apps.

Have it pulling content now, and cleaning up the content. About ready to
integrate S3 (Minio) and do some testing of object storage, and then
packing into DBs.